orcid preferences based on hidden/selected relations
for the orcid profiles owner with OrcidEntitySyncPreference MINE and MY_SELECTED for some entity an additional check was added.
The check uses the solr-search to determine, if the corresponding/linked item is among the hidden/preferred items of the orcid profile owner.
The solr query for the relation can be configured.