save additional value in Controlled Vocabulary
some optional label can be added to controlledvocabularies which is saved when given. when the value is not given there is the fallback that the id can be saved (which must be unique). This allows parent nodes to have the same saved vakze as child noses, although they have some different id.